TextArenaMatchmakingServer/requirements.txt
colorlog
supabase
fastapi
uvicorn
python-dotenv
numpy
boto3
requests
tabulate
textarena